+++ variables.tf
@@ -79,6 +79,12 @@ variable "apiary_log_prefix" {
   default     = ""
 }
 
+variable "enable_hive_metastore_metrics" {
+  description = "Enable sending Hive Metastore metrics to CloudWatch."
+  type        = "string"
+  default     = ""
+}
+
 variable "apiary_managed_schemas" {
   description = "Schema names from which S3 bucket names will be derived, corresponding S3 bucket will be named as apiary_instance-aws_account-aws_region-schema_name."
   type        = "list"